requirements.txt
mysql-connector-python
numpy
python-dateutil
argparse
requests
configparser